Modification du LaunchFIle pour qu'il prenne maintenant le nouveau splashscreen Mise a jour de la ProfileVC pour répondre aux besoins de l'app Ajout des images du Launchfile Ajout des images pour l'icône

+++ Frismart/Sources/Core/STHelpers.swift
@@ -46,7 +46,7 @@ class STHelpers: NSObject {
     
     class func getNearestStore()->STStore {
     
-        var distance: Double = Double(99999999999)
+        var distance: Double = Double(999999)
         var newDistance: Double = Double(0)
         var nearestStore: STStore = STStore()
